--- src/main/java/com/anyang/maruni/domain/llm/application/port/LlmClient.java
@@ -1,5 +1,0 @@
-package com.anyang.maruni.domain.llm.application.port;
-
-public interface LlmClient {
-    String chat(String prompt);
-}
