--- src/main/java/com/anyang/maruni/domain/stt/application/port/SttClient.java
@@ -1,7 +1,0 @@
-package com.anyang.maruni.domain.stt.application.port;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface SttClient {
-    String transcribe(MultipartFile audioFile);
-}
