--- src/main/java/com/anyang/maruni/domain/stt/domain/repository/ConversationRepository.java
@@ -1,7 +1,0 @@
-package com.anyang.maruni.domain.stt.domain.repository;
-
-import com.anyang.maruni.domain.stt.domain.entity.Conversation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ConversationRepository extends JpaRepository<Conversation, Long> {
-}
