--- src/main/java/com/anyang/maruni/domain/tts/application/port/TtsClient.java
@@ -1,5 +1,0 @@
-package com.anyang.maruni.domain.tts.application.port;
-
-public interface TtsClient {
-    byte[] synthesizeSpeech(String text);
-}
