--- src/main/java/com/anyang/maruni/domain/tts/infra/TtsApiClient.java
@@ -1,31 +1,0 @@
-package com.anyang.maruni.domain.tts.infra;
-
-import com.anyang.maruni.domain.tts.application.port.TtsClient;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import java.util.Map;
-
-@Component
-@RequiredArgsConstructor
-public class TtsApiClient implements TtsClient {
-
-    private final WebClient openAiWebClient;
-
-    @Override
-    public byte[] synthesizeSpeech(String text) {
-
-        return openAiWebClient.post()
-                .uri("/audio/speech")
-                .bodyValue(Map.of(
-                        "model", "tts-1",
-                        "input", text,
-                        "voice", "nova",
-                        "speed", 1.0
-                ))
-                .retrieve()
-                .bodyToMono(byte[].class)  // binary response
-                .block();
-    }
-}
