+++ src/main/java/com/anyang/maruni/domain/voice_chat/application/port/LlmClient.java
@@ -1,0 +1,5 @@
+package com.anyang.maruni.domain.voice_chat.application.port;
+
+public interface LlmClient {
+    String chat(String prompt);
+}
