+++ src/main/java/com/anyang/maruni/domain/voice_chat/application/port/TtsClient.java
@@ -1,0 +1,5 @@
+package com.anyang.maruni.domain.voice_chat.application.port;
+
+public interface TtsClient {
+    byte[] synthesizeSpeech(String text);
+}
