+++ src/main/java/com/anyang/maruni/domain/voice_chat/domain/repository/ConversationRepository.java
@@ -1,0 +1,7 @@
+package com.anyang.maruni.domain.voice_chat.domain.repository;
+
+import com.anyang.maruni.domain.voice_chat.domain.entity.Conversation;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ConversationRepository extends JpaRepository<Conversation, Long> {
+}
